backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.0
ephem==4.2
fastapi==0.115.12
h11==0.16.0
idna==3.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
setuptools==80.4.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
wheel==0.45.1
flake8==7.0.0
black==24.2.0
requests==2.31.0
python-dotenv==1.0.1
